New requests supported
* Fixed imports
* Add testing dependencies
* Add Check* Family
* Add Read* Family
* Renamed package
