+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/MainRunner.java
@@ -1,15 +1,20 @@
 package edu.finplatjavacourse.distributeddbprototype;
 
 
-import edu.finplatjavacourse.distributeddbprototype.request.DelegatingRequestHandler;
-import edu.finplatjavacourse.distributeddbprototype.request.RequestHandler;
-import edu.finplatjavacourse.distributeddbprototype.request.executor.StatementExecutor;
-import edu.finplatjavacourse.distributeddbprototype.request.executor.WriteStatementExecutor;
-import edu.finplatjavacourse.distributeddbprototype.request.parsing.RequestParser;
-import edu.finplatjavacourse.distributeddbprototype.request.parsing.impl.WriteRequestParser;
-import edu.finplatjavacourse.distributeddbprototype.request.response.AlreadyExistsResponseTransformer;
-import edu.finplatjavacourse.distributeddbprototype.request.response.ResponseTransformer;
-import edu.finplatjavacourse.distributeddbprototype.request.response.SimpleResponseTransformer;
+import edu.finplatjavacourse.distributeddbprototype.handler.DelegatingRequestHandler;
+import edu.finplatjavacourse.distributeddbprototype.handler.RequestHandler;
+import edu.finplatjavacourse.distributeddbprototype.handler.executor.CheckStatementExecutor;
+import edu.finplatjavacourse.distributeddbprototype.handler.executor.ReadStatementExecutor;
+import edu.finplatjavacourse.distributeddbprototype.handler.executor.StatementExecutor;
+import edu.finplatjavacourse.distributeddbprototype.handler.executor.WriteStatementExecutor;
+import edu.finplatjavacourse.distributeddbprototype.handler.parsing.RequestParser;
+import edu.finplatjavacourse.distributeddbprototype.handler.parsing.impl.CheckRequestParser;
+import edu.finplatjavacourse.distributeddbprototype.handler.parsing.impl.ReadRequestParser;
+import edu.finplatjavacourse.distributeddbprototype.handler.parsing.impl.WriteRequestParser;
+import edu.finplatjavacourse.distributeddbprototype.handler.response.AlreadyExistsResponseTransformer;
+import edu.finplatjavacourse.distributeddbprototype.handler.response.ReadResponseTransformer;
+import edu.finplatjavacourse.distributeddbprototype.handler.response.ResponseTransformer;
+import edu.finplatjavacourse.distributeddbprototype.handler.response.SimpleResponseTransformer;
 
 
 public class MainRunner {
@@ -19,27 +24,43 @@ public class MainRunner {
                 statementExecutorChain(),
                 responseTransformerChain());
 
+        System.out.println("=======\"PUT 123, \\\"Hotel 123 \\\", 100\"=========");
         System.out.println(requestHandler.handle("PUT 123, \"Hotel 123 \", 100")); // should be valid
+        System.out.println("=======\"PUT 321, \\\"Hotel 321 \\\", 100\"=========");
         System.out.println(requestHandler.handle("PUT 321, \"Hotel 321 \", 100")); // should be valid
 
+        System.out.println("=======\"PUT 123, \\\"Hotel 123 \\\", 100\"=========");
         System.out.println(requestHandler.handle("PUT 123, \"Hotel 123 \", 100")); // should be invalid since such id already exists
 
+        System.out.println("=======\"DFSFDSF\"=========");
         System.out.println(requestHandler.handle("DFSFDSF")); // should be invalid since bad validated
+        System.out.println("=======\"PUT e234, \\\"Hotel 123 \\\", 100\"=========");
         System.out.println(requestHandler.handle("PUT e234, \"Hotel 123 \", 100")); // should be invalid since bad validated
+        System.out.println("=======\"PUT 100, \\\"Hotel 123 \\\", dsafa54\"=========");
         System.out.println(requestHandler.handle("PUT 100, \"Hotel 123 \", dsafa54")); // should be invalid since bad validated
+
+        System.out.println("=======\"GET \\\"Hotel\\\"\"=========");
+        System.out.println(requestHandler.handle("GET \"Hotel\"")); // should be valid
+        System.out.println("=======\"GET adsfasdfasdf\"=========");
+        System.out.println(requestHandler.handle("GET adsfasdfasdf")); // should be invalid
+
+        System.out.println("=======\"CHECK PUT 321, \\\"Hotel 321\\\", 100\"=========");
+        System.out.println(requestHandler.handle("CHECK PUT 321, \"Hotel 321\", 100")); // should be valid
+        System.out.println("=======\"CHECK GET adsfasdfasdf\"=========");
+        System.out.println(requestHandler.handle("CHECK GET adsfasdfasdf")); // should be invalid
+        System.out.println("=======\"CHECK PUT adsfasdfasdf\"=========");
+        System.out.println(requestHandler.handle("CHECK PUT adsfasdfasdf")); // should be invalid
     }
 
     private static RequestParser requestParserChain() {
-        return new WriteRequestParser();
+        return new ReadRequestParser(new CheckRequestParser(new WriteRequestParser()));
     }
 
     private static StatementExecutor statementExecutorChain() {
-        return new WriteStatementExecutor();
+        return new ReadStatementExecutor(new CheckStatementExecutor(new WriteStatementExecutor()));
     }
 
     private static ResponseTransformer responseTransformerChain() {
-        return new SimpleResponseTransformer(
-                new AlreadyExistsResponseTransformer()
-        );
+        return new ReadResponseTransformer(new SimpleResponseTransformer(new AlreadyExistsResponseTransformer()));
     }
 }
