+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/handler/RequestHandler.java
@@ -1,0 +1,8 @@
+package edu.finplatjavacourse.distributeddbprototype.handler;
+
+
+public interface RequestHandler {
+
+    String handle(String request);
+
+}
