+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/handler/exception/ExecutionException.java
@@ -1,0 +1,12 @@
+package edu.finplatjavacourse.distributeddbprototype.handler.exception;
+
+
+public class ExecutionException extends StatementProcessingException {
+    public ExecutionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ExecutionException(String message) {
+        super(message);
+    }
+}
