+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/handler/exception/ParsingException.java
@@ -1,0 +1,12 @@
+package edu.finplatjavacourse.distributeddbprototype.handler.exception;
+
+
+public class ParsingException extends RequestHandlingException {
+    public ParsingException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ParsingException(String message) {
+        super(message);
+    }
+}
