+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/handler/exception/ResponseTransformerException.java
@@ -1,0 +1,12 @@
+package edu.finplatjavacourse.distributeddbprototype.handler.exception;
+
+
+public class ResponseTransformerException extends RequestHandlingException {
+    public ResponseTransformerException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ResponseTransformerException(String message) {
+        super(message);
+    }
+}
