+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/handler/exception/StatementProcessingException.java
@@ -1,0 +1,12 @@
+package edu.finplatjavacourse.distributeddbprototype.handler.exception;
+
+
+public class StatementProcessingException extends RequestHandlingException {
+    public StatementProcessingException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public StatementProcessingException(String message) {
+        super(message);
+    }
+}
