+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/handler/parsing/RequestToStatementTransformer.java
@@ -1,0 +1,8 @@
+package edu.finplatjavacourse.distributeddbprototype.handler.parsing;
+
+
+public interface RequestToStatementTransformer {
+
+    Statement transform(String request);
+
+}
