+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/handler/parsing/RequestValidator.java
@@ -1,0 +1,8 @@
+package edu.finplatjavacourse.distributeddbprototype.handler.parsing;
+
+
+public interface RequestValidator {
+    void validate(String request);
+
+    boolean canParse(String request);
+}
