+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/handler/parsing/Statement.java
@@ -1,0 +1,5 @@
+package edu.finplatjavacourse.distributeddbprototype.handler.parsing;
+
+
+public interface Statement {
+}
