+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/handler/parsing/impl/ReadStatement.java
@@ -1,0 +1,8 @@
+package edu.finplatjavacourse.distributeddbprototype.handler.parsing.impl;
+
+
+import edu.finplatjavacourse.distributeddbprototype.handler.parsing.Statement;
+
+
+public record ReadStatement(String searchOn) implements Statement {
+}
