+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/handler/parsing/impl/WriteStatement.java
@@ -1,0 +1,8 @@
+package edu.finplatjavacourse.distributeddbprototype.handler.parsing.impl;
+
+
+import edu.finplatjavacourse.distributeddbprototype.handler.parsing.Statement;
+
+
+public record WriteStatement(long id, String name, long price) implements Statement {
+}
