+++ src/main/java/edu/finplatjavacourse/distributeddbprototype/handler/response/SimpleResponseTransformer.java
@@ -1,0 +1,22 @@
+package edu.finplatjavacourse.distributeddbprototype.handler.response;
+
+
+public class SimpleResponseTransformer extends ResponseTransformer {
+    public SimpleResponseTransformer(ResponseTransformer next) {
+        super(next);
+    }
+
+    public SimpleResponseTransformer() {
+        super(null);
+    }
+
+    @Override
+    protected String transform0(Response response) {
+        return ((Response.SimpleResponse) response).ok() ? "OK" : "FAIL";
+    }
+
+    @Override
+    protected boolean canTransform(Response response) {
+        return response instanceof Response.SimpleResponse;
+    }
+}
