--- src/main/java/edu/finplatjavacourse/distributeddbprototype/request/RequestHandler.java
@@ -1,8 +1,0 @@
-package edu.finplatjavacourse.distributeddbprototype.request;
-
-
-public interface RequestHandler {
-
-    String handle(String request);
-
-}
