--- src/main/java/edu/finplatjavacourse/distributeddbprototype/request/exception/ExecutionException.java
@@ -1,12 +1,0 @@
-package edu.finplatjavacourse.distributeddbprototype.request.exception;
-
-
-public class ExecutionException extends StatementProcessingException {
-    public ExecutionException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ExecutionException(String message) {
-        super(message);
-    }
-}
