--- src/main/java/edu/finplatjavacourse/distributeddbprototype/request/exception/ParsingException.java
@@ -1,12 +1,0 @@
-package edu.finplatjavacourse.distributeddbprototype.request.exception;
-
-
-public class ParsingException extends RequestHandlingException {
-    public ParsingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ParsingException(String message) {
-        super(message);
-    }
-}
