--- src/main/java/edu/finplatjavacourse/distributeddbprototype/request/exception/ResponseTransformerException.java
@@ -1,12 +1,0 @@
-package edu.finplatjavacourse.distributeddbprototype.request.exception;
-
-
-public class ResponseTransformerException extends RequestHandlingException {
-    public ResponseTransformerException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ResponseTransformerException(String message) {
-        super(message);
-    }
-}
