--- src/main/java/edu/finplatjavacourse/distributeddbprototype/request/exception/StatementProcessingException.java
@@ -1,12 +1,0 @@
-package edu.finplatjavacourse.distributeddbprototype.request.exception;
-
-
-public class StatementProcessingException extends RequestHandlingException {
-    public StatementProcessingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public StatementProcessingException(String message) {
-        super(message);
-    }
-}
