--- src/main/java/edu/finplatjavacourse/distributeddbprototype/request/exception/ValidatingException.java
@@ -1,12 +1,0 @@
-package edu.finplatjavacourse.distributeddbprototype.request.exception;
-
-
-public class ValidatingException extends ParsingException {
-    public ValidatingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ValidatingException(String message) {
-        super(message);
-    }
-}
