--- src/main/java/edu/finplatjavacourse/distributeddbprototype/request/parsing/RequestToStatementTransformer.java
@@ -1,8 +1,0 @@
-package edu.finplatjavacourse.distributeddbprototype.request.parsing;
-
-
-public interface RequestToStatementTransformer {
-
-    Statement transform(String request);
-
-}
