--- src/main/java/edu/finplatjavacourse/distributeddbprototype/request/parsing/RequestValidator.java
@@ -1,8 +1,0 @@
-package edu.finplatjavacourse.distributeddbprototype.request.parsing;
-
-
-public interface RequestValidator {
-    void validate(String request);
-
-    boolean canParse(String request);
-}
