--- src/main/java/edu/finplatjavacourse/distributeddbprototype/request/parsing/Statement.java
@@ -1,5 +1,0 @@
-package edu.finplatjavacourse.distributeddbprototype.request.parsing;
-
-
-public interface Statement {
-}
